requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
Django==4.2.10
djangorestframework==3.14.0
psycopg==3.1.18
pytz==2024.1
sqlparse==0.4.4
typing_extensions==4.9.0
django-cors-headers==4.3.1
python-decouple==3.8
django-jazzmin==2.6.0
djangorestframework-simplejwt==5.3.1
